requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
fastapi>=0.111.0
uvicorn>=0.30.0
python-dateutil>=2.9.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
exchange-calendars>=4.13.2
